Change the LicenseDialog into an AcceptDialog.
Changing the LicenseDialog from a Window into an AcceptDialog allows me
to configure it so that the window closes when clicked out of, or when
the Escape key is pressed.  It also provides the user with an obvious OK
button.  This is handy if, for example, the export doesn't render the
window close button.