Standardize i18n labels.
Most of the keys in the localization file now conform to certain
standards.  The desired effect is that most labels which serve a common
purpose will appear in a common namespace, especially important if the
user is sorting the spreadsheet by keys.

Menu button labels now start with "MENU_BUTTON_" (as in the example of
the debug menu button label from the previous commit) and labels
relevant to the quit window now start with "QUIT_WINDOW_".

An effort will be made to maintain this naming scheme into the future.